requirements.txt
torch
openai-clip
opencv-python
torchvision
scikit-learn
matplotlib==3.7.3
pandas
seaborn
grad-cam
lime
shap
tqdm
quantus
transformers==4.34.0 
einops
bitsandbytes
sentencepiece
accelerate
cachetools
pillow
open-clip-torch
fairscale
tensorflow
xplique